♻ refactor(#6): home → planet 메인 경로명 수정
